files
This project aims to develop a web automation tool for active sales requests using technologies such as web scraping, ReactJS, Django, PostgreSQL, and Python
scripting. The tool will automate the process of finding the best page for the customer
based on their required cost, thereby reducing the time and effort required to search
for the best deal manually.